Deduplicate autoloop pipelines: extract shared steps
- Move hp-adjust-ps and metric-decision-ps to shared-autoloop/ (single
  canonical copy, previously duplicated across all 3 autoloop pipelines)
- Symlink from each autoloop pipeline to shared-autoloop/
- Consolidate 6 task-specific test files into 2 unified suites (106 tests)
- Fix sys.modules collision by using importlib for test imports
- Move retune-pipeline-design docs into shared-autoloop/
