+++ detector-pipeline-yolof-autoloop/hp-adjust-ps
@@ -1,0 +1,1 @@
+../shared-autoloop/hp-adjust-ps
