--- detector-pipeline-yolof-autoloop/hp-adjust-ps/1/pipeline_step.py
@@ -1,15 +1,0 @@
-#!/usr/bin/env python
-import sys
-from pathlib import Path
-
-sys.path.insert(0, str(Path(__file__).parent / "models"))
-model_module = __import__("model.1.model", fromlist=[''])
-model_class = [obj for name in dir(model_module) if isinstance(obj := getattr(model_module, name), type) and hasattr(obj, 'adjust')][0]
-
-def main():
-    args = model_class.to_pipeline_parser().parse_args()
-    model_class().adjust(**vars(args))
-
-
-if __name__ == "__main__":
-    main()
