--- lora-pipeline-unsloth-autoloop/hp-adjust-ps/Dockerfile
@@ -1,10 +1,0 @@
-FROM --platform=linux/amd64 python:3.11-slim
-
-ENV DEBIAN_FRONTEND=noninteractive
-ENV TZ=UTC
-
-COPY --link requirements.txt /home/nonroot/requirements.txt
-RUN pip install --no-cache-dir -r /home/nonroot/requirements.txt
-
-COPY --link=true 1 /home/nonroot/main/1
-COPY --link=true requirements.txt config.yaml /home/nonroot/main/
