--- lora-pipeline-unsloth-autoloop/hp-adjust-ps/requirements.txt
@@ -1,1 +1,0 @@
-clarifai>=12.1.4,<13.0.0
