+++ shared-autoloop/hp-adjust-ps/Dockerfile
