+++ shared-autoloop/hp-adjust-ps/config.yaml
