+++ shared-autoloop/retune-pipeline-design/hp-adjustment-design.md
